pyproject.toml
[tool.poetry]
name = "cryptopals"
version = "0.1.0"
description = ""
authors = ["Martyn <martyn.smith560@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycryptodome = "^3.10"

[tool.pytest.ini_options]
minversion = "6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
